misc: add validation and preview return type

+++ RentingCarAPI/Controllers/VehicleTypeController.cs
@@ -1,5 +1,6 @@
 ﻿using BusinessObjects.Models;
 using Microsoft.AspNetCore.Mvc;
+using RentingCarAPI.ViewModel;
 using RentingCarServices.ServiceInterface;
 
 namespace RentingCarAPI.Controllers
@@ -16,36 +17,68 @@ namespace RentingCarAPI.Controllers
             _logger = logger;
             _vehicleTypeService = vehicleTypeService;
         }
-        [HttpGet("getAllVehicleTypes")]
+        [HttpGet("getall")]
+        [ProducesResponseType(typeof(List<VehicleType>), StatusCodes.Status200OK)]
+        [ProducesResponseType(typeof(ResponseVM), StatusCodes.Status404NotFound)]
         public IActionResult GetAll()
         {
             List<VehicleType> typeList = _vehicleTypeService.GetVehicleTypes();
             if (!typeList.Any())
             {
-                return NotFound("There is no vehicle type");
+                return NotFound(new ResponseVM { 
+                    Message = "There's No Vehicle List",
+                    Errors = new string[] {"Vehicle Type List is Empty"}
+                });
             }
             return Ok(typeList);
         }
-        [HttpPost("createType")]
+        [HttpPost("create")]
+        [ProducesResponseType(typeof(ResponseVMWithEntity<VehicleType>), StatusCodes.Status200OK)]
+        [ProducesResponseType(typeof(ResponseVM), StatusCodes.Status400BadRequest)]
+        [ProducesResponseType(typeof(ResponseVMWithEntity<VehicleType>), StatusCodes.Status400BadRequest)]
         public IActionResult Create(string name)
         {
             if(!ModelState.IsValid)
             {
-                return BadRequest();
+                return BadRequest(new ResponseVM
+                {
+                    Message = "Invalid Input",
+                    Errors = new string[] {"Name is null", "Name must have at least 50 characters"}
+                });
             }
             VehicleType newVehicleType = new VehicleType();
             newVehicleType.TypeName = name;
             bool check = _vehicleTypeService.Add(newVehicleType);
             if(!check)
             {
-                return BadRequest();
+                return BadRequest(new ResponseVMWithEntity<VehicleType>
+                {
+                    Message = "Invalid Input",
+                    Entity = newVehicleType,
+                    Errors = new string[] {"Invalid Data to Database", "Cannot Save or Update to Database"}
+                });
             }
 
-            return Ok("Create Successfully\n" +newVehicleType);
+            return Ok(new ResponseVMWithEntity<VehicleType>
+            {
+                Message = "Create Successfully",
+                Entity = newVehicleType
+            });
         }
-        [HttpPut("updateType/{id}")]
-        public IActionResult Update([FromRoute] long id,string name)
+        [HttpPut("update/{id}")]
+        [ProducesResponseType(typeof(ResponseVMWithEntity<VehicleType>), StatusCodes.Status200OK)]
+        [ProducesResponseType(typeof(ResponseVM), StatusCodes.Status400BadRequest)]
+        [ProducesResponseType(typeof(ResponseVMWithEntity<VehicleType>), StatusCodes.Status400BadRequest)]
+        public ActionResult<VehicleType> Update([FromRoute] long id,string name)
         {
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(new ResponseVM
+                {
+                    Message = "Invalid Data",
+                    Errors = new string[] { "Name is null", "Name must have at least 50 characters" }
+                });
+            }
             VehicleType updateType = _vehicleTypeService.GetVehicleTypeById(id);
             if (updateType != null)
             {
@@ -54,19 +87,48 @@ namespace RentingCarAPI.Controllers
             bool check = _vehicleTypeService.Update(updateType);
             if(!check)
             {
-                return BadRequest();
+                return BadRequest(new ResponseVMWithEntity<VehicleType>
+                {
+                    Message = "Invalid Input",
+                    Entity = updateType,
+                    Errors = new string[] {"Invalid Data to Database", "Cannot Save or Update to Database" }
+                });
             }
-            return Ok("Update Successfully");
+            return Ok(new ResponseVMWithEntity<VehicleType>
+            {
+                Message = "Update Successfully",
+                Entity = updateType
+            });
         }
-        [HttpDelete("deleteType/{id}")]
+        [HttpDelete("delete/{id}")]
+        [ProducesResponseType(typeof(ResponseVMWithEntity<VehicleType>), StatusCodes.Status200OK)]
+        [ProducesResponseType(typeof(ResponseVM), StatusCodes.Status400BadRequest)]
+        [ProducesResponseType(typeof(ResponseVMWithEntity<VehicleType>), StatusCodes.Status404NotFound)]
         public IActionResult Delete([FromRoute] long id) {
             VehicleType delType = _vehicleTypeService.GetVehicleTypeById(id);
+            if (delType == null)
+            {
+                return NotFound(new ResponseVM
+                {
+                    Message = "Cannot find type with id" +id,
+                    Errors = new string[] {"There's no data in database with id" +id}
+                });
+            }
             bool check = _vehicleTypeService.Delete(delType);
             if (!check)
             {
-                return BadRequest();
+                return BadRequest(new ResponseVMWithEntity<VehicleType>
+                {
+                    Message = "Invalid Data",
+                    Entity = delType,
+                    Errors = new string[] {"Cannot Save or Update to Database", "Invalid Data to Database"}
+                });
             }
-            return Ok("Delete Successfully Type With ID: " +id);
+            return Ok(new ResponseVMWithEntity<VehicleType>
+            {
+                Message = "Delete Successfully",
+                Entity = delType
+            });
         }
     }
 }
